+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* New
+  * When adding session, pre-select game with the last session
+
 ## 0.2.4 / 2023-01-16 19:39+01:00
 
 * Fixed